fix: esm full path support and simplified bins
