+++ .esm.swcrc
@@ -8,6 +8,7 @@
         "experimental": {
             "plugins": [["@hanseltime/swc-plugin-node-globals-inject", {}]]
         },
+        "baseUrl": "."
     },
     // Since we use tswc most of the other options are mapped from tsconfig.  
     // We just override differences or exotic properties here
